# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel CQRS",
        "version": "v1"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "description": "Get all users list",
                "operationId": "32f813ea7a22e2b8b862ef93084426b9",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "Adds a new user",
                "operationId": "c70c18636cb91d835fcc23025f48febe",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/users/getById": {
            "get": {
                "tags": [
                    "Users"
                ],
                "description": "Get user by id",
                "operationId": "024e66a04a45d7a2878f0b0cf1bca1db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/users/{id}/update": {
            "put": {
                "tags": [
                    "Users"
                ],
                "description": "Update user by id",
                "operationId": "92d9d95c96cc4c903b463d2dd28ee6f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    }
}